refactor: TrendingArtists section
